move minio hook to a new module and imporve it
